factory

+++ src/msys/core/processor.py
@@ -10,12 +10,13 @@ from typing import Optional
 class Processor:
     def __init__(self,
                  id,
-                 launch,
+                 location: Optional[str] = None,
+                 remote: Optional[str] = "",
                  users: Optional[int]=0):
         self.id = id
-        self.launch = launch
+        self.launch = location
         self.process = None
-        self.url = "http://127.0.0.1:{port}"
+        self.url = remote
         self.port = -1
         self.users = users
 
@@ -32,15 +33,16 @@ class Processor:
     def start(self):
         self.increase_users()
 
-        if self.users > 0 or self.started:
+        if self.users > 1 or self.started:
             return
 
-        if type(self.launch) == str:
-            if self.launch.find("{port}") != -1:
-                self.port = find_open_ports()
-                self.url = "http://127.0.0.1:{port}".replace("{port}", str(self.port))
-                cmd = self.launch.replace("{port}", str(self.port))
-                self.process = sp.Popen(cmd)
+        if self.launch:
+            if type(self.launch) == str:
+                if self.launch.find("{port}") != -1:
+                    self.port = find_open_ports()
+                    self.url = "http://127.0.0.1:{port}".replace("{port}", str(self.port))
+                    cmd = self.launch.replace("{port}", str(self.port))
+                    self.process = sp.Popen(cmd)
 
         self.started = True
 
@@ -53,11 +55,13 @@ class Processor:
         self.kill()
 
     def kill(self):
-        ptype = type(self.process)
-        if ptype == sp.Popen:
-            self.process.terminate()
-        else:
-            print("[Process] dont found: " + str(ptype))
+        print("[Process] will be killed: " + self.url)
+        if self.launch:
+            ptype = type(self.process)
+            if ptype == sp.Popen:
+                self.process.terminate()
+            else:
+                print("[Process] dont found: " + str(ptype))
 
         self.started = False
 
@@ -68,6 +72,7 @@ class Processor:
             if response.status_code != 200:
                 return dict()
             return json.loads(response.content)
+        print("[Processor] no url")
 
     def change_config(self, config:dict) -> dict:
         if self.url:
@@ -75,7 +80,7 @@ class Processor:
             if response.status_code != 200:
                 return dict()
             return json.loads(response.content)
-
+        print("[Processor] no url")
 
     def update_config(self, config:dict) -> dict:
         if self.url:
@@ -83,3 +88,7 @@ class Processor:
             if response.status_code != 200:
                 return dict()
             return json.loads(response.content)
+        print("[Processor] no url")
+
+    def __del__(self):
+        self.kill()
